feat: Display local plagiarism results
The `url` property of the plagiarism DTO contains a potentially local
URL of the result, while the `result` property contains the response of
the plagiarism service, which is an implementation detail, and always
contains the remote URL of the Moss page (but it can be something
completely different if more plagiarism check services are supported in
the future). Also document that it can be `null` (but not `undefined`)
if the plagiarism check has not yet been run (or it was unsuccessful).